@@ -5,20 +5,20 @@ RSpec.describe Admin::ContentsController, type: :controller do
 
   describe "functionalities with logged in user with role 'admin'" do
     before do
-      user = FactoryGirl.create(:user_with_admin_role)
+      user = FactoryBot.create(:user_with_admin_role)
       sign_in user
     end
     # This should return the minimal set of attributes required to create a valid
     # Content. As you add validations to Content, be sure to
     # adjust the attributes here as well.
     let(:content) {
-      FactoryGirl.create(:content)
+      FactoryBot.create(:content)
     }
     let(:valid_attributes) {
-      FactoryGirl.build(:content).attributes
+      FactoryBot.build(:content).attributes
     }
     let(:invalid_attributes) {
-      FactoryGirl.attributes_for(:content, :invalid)
+      FactoryBot.attributes_for(:content, :invalid)
     }
 
     # This should return the minimal set of values that should be in the session
@@ -93,7 +93,7 @@ RSpec.describe Admin::ContentsController, type: :controller do
     describe "PUT #update" do
       context "with valid params" do
         let(:new_attributes) {
-          FactoryGirl.attributes_for(:content, title: "new content title")
+          FactoryBot.attributes_for(:content, title: "new content title")
 
         }
 
