@@ -8,18 +8,18 @@ RSpec.describe Admin::LocationsController, type: :controller do
   describe "functionalities with logged in user with role 'admin'" do
     before do
       @request.env['devise.mapping'] = Devise.mappings[:user]
-      user = FactoryGirl.create(:user_with_admin_role)
+      user = FactoryBot.create(:user_with_admin_role)
       sign_in user
     end
 
     let(:location) {
-      FactoryGirl.create(:location)
+      FactoryBot.create(:location)
     }
     let(:valid_attributes) {
-      FactoryGirl.build(:location).attributes
+      FactoryBot.build(:location).attributes
     }
     let(:invalid_attributes) {
-      FactoryGirl.attributes_for(:location, :invalid)
+      FactoryBot.attributes_for(:location, :invalid)
     }
 
     # This should return the minimal set of values that should be in the session
@@ -94,7 +94,7 @@ RSpec.describe Admin::LocationsController, type: :controller do
     describe "PUT #update" do
       context "with valid params" do
         let(:new_attributes) {
-          FactoryGirl.attributes_for(:location, title: "new location title")
+          FactoryBot.attributes_for(:location, title: "new location title")
         }
 
         it "updates the requested location" do
