@@ -1,4 +1,4 @@
-FactoryGirl.define do
+FactoryBot.define do
   factory :flyer do
     image { Rack::Test::UploadedFile.new(File.join(Rails.root, 'spec', 'support', 'flyer.jpg'), 'image/jpg') }
     title "MyString"
