@@ -1,4 +1,4 @@
-FactoryGirl.define do
+FactoryBot.define do
 
   sequence :login do |login|
     "#{Faker::Internet.user_name}#{login}".gsub(/[^a-z0-9]/, "-")
@@ -21,7 +21,7 @@ end
 
 # users with role
 def define_user_with_role(role)
-  FactoryGirl.define do
+  FactoryBot.define do
 
     factory :"user_with_#{role}_role", :parent => :user do |u|
       u.login { generate :login }
