@@ -9,15 +9,15 @@ RSpec.describe "Programs", type: :request do
 
   describe "GET /programs" do
     it "redirects to another program if it is defined as a current_program" do
-      @program = FactoryGirl.create(:program)
-      @another_program = FactoryGirl.create(:program, :current, :title => "Another program")
+      @program = FactoryBot.create(:program)
+      @another_program = FactoryBot.create(:program, :current, :title => "Another program")
       get programs_path
       expect(response).to have_http_status(302)
       expect(response).to redirect_to(program_path(@helpers.current_program))
     end
 
     it "redirects to this program if it is defined as the current_program" do
-      @program = FactoryGirl.create(:program, :current)
+      @program = FactoryBot.create(:program, :current)
       allow(@helpers).to receive(:current_program).and_return(@program)
       get programs_path
       expect(response).to have_http_status(302)
@@ -25,7 +25,7 @@ RSpec.describe "Programs", type: :request do
     end
 
     it "redirects to root url if a current_program is NOT defined" do
-      @program = FactoryGirl.create(:program)
+      @program = FactoryBot.create(:program)
       get programs_path
       expect(response).to have_http_status(302)
       # fix me in helpers/application_helper.rb:
