@@ -1,7 +1,7 @@
 class ApplicationController < ActionController::Base
 
   helper :all
-  protect_from_forgery
+  protect_from_forgery with: :exception
 
 
   def after_sign_in_path(resource)
