@@ -20,4 +20,5 @@ ActiveSupport.to_time_preserves_timezone = false
 Rails.application.config.active_record.belongs_to_required_by_default = false
 
 # Do not halt callback chains when a callback returns false. Previous versions had true.
-ActiveSupport.halt_callback_chains_on_return_false = true
+# deprecated at rails 5.1
+# ActiveSupport.halt_callback_chains_on_return_false = true
